Criação da interface gráfica e integração com os casos de usos criados

+++ lib/app/modules/home/domain/repositories_interfaces/number_trivia_repository_interface.dart
@@ -3,5 +3,5 @@ import 'package:trivia_app_flutter/app/modules/home/infra/models/trivia_model.da
 abstract class INumberTriviaRepository {
 
 
-  Future<TriviaModel> getTrivia({int numberTrivia});
+  Future<TriviaModel> getTrivia({double numberTrivia});
 }
