+++ lib/app/modules/home/infra/external_interfaces/data_souces_interfaces/number_trivia_data_source_interface.dart
@@ -1,5 +1,5 @@
 abstract class INumberTriviaDataSource{
 
-  Future<Map<String, dynamic>> getTriviaData({int numberTrivia});
+  Future<Map<String, dynamic>> getTriviaData({double numberTrivia});
 
 }
